Add colors file to kvGlowDark theme
Add colors file to kvGlowDark theme.  Without this QTableView and
QListView have ugly background colors when disabled. The default
palette colors used in the .kvconfig file don't actually cater for
the disabled state of widgets. So it looks like we need to use the
colors file.

The kvGlowDark.colors is a copy of DarkLines.colors.