Refactor the theme setting from style setting in test app on launch
Refactor the theme setting from style setting in test app on launch.
It seems that in order get the *style engine* and the theme fully
applied on each widget the setup must be in two steps. First create
the style engine and apply whatever it does before window is opened
then apply the theme (and set the palette) after the main window is
open.

If the style is created and the theme is loaded in the style plugin
style create function widgets have the right look, shape etc. but the
colors from the .colors file are not applied. When the theme is loaded
after the window is open and the window is restyled then the .colors
take effect. However if the style engine was not created with the matching
theme setup (before the window) open some widgets will have wrong look

I don't know what is the fundamental issue but I'm guessing it's got
something to do with some persistent state being stored somwhere,
probably in the widgets.